refactor: split visibility purge job registration into its own OSGi component
Separate the Quartz-executed job (VisibilityActionPurgeJob) from its OSGi
registration (VisibilityActionPurgeJobRegistration), following the Jahia
background-job sample pattern. The job class is no longer an OSGi component
since Quartz instantiates it directly without dependency injection.
